+++ higepos/src/include/higeOperator.h
@@ -12,10 +12,12 @@
     \date   create:2002/08/08 update:$Date$
 */
 
+#include<higepostypes.h>
+
 #ifndef _HIGE_POS_OPERATOR
 #define _HIGE_POS_OPERATOR
 
-void* operator new(unsigned int size);
+void* operator new(size_t size);
 void  operator delete(void* address);
 void* malloc(unsigned long size);
 void __builtin_delete(void* address);
